Redid grass path handling logic to be more detached
Got rid of custom grass path block that overwrote the base grass path block in the block registry. In exchange added an event handler to catch when a grass path block is updated by its' neighbors (such as a fence gate being placed on it) so that it would not revert to a dirt block. Also had to add a bunch of code copied from World and Chunk to update their blockstate info and related stuff to allow a call to grass path's onBlockAdded method which caused the block to revert to dirt.